Updated Docker Image
The old docker image (node:latest) is not best practice as it

a.) Doesn't pin the NodeJS version. If NodeJS were to update, we would automatically update along with it. If there were some sort of breaking change to a newer version of NodeJS, our application might stop working.

b.) `node:latest` is a large image. Once changed to `node:13.8-alpine`, the image shrunk from 1.13GB to 305MB. about 1/3 of the original size.