Improvements to the Alias System
Renamed all the new variable from command/s to alias/es as it aligns more with the update name, as recommended by c650.
Plugins no longer are required to have the alias section.
When the code detects an duplicate aliases it will send a warning on the plugins that collide, olong with the action it took with this problem.