Implemented a sub plugin, that implements Abes new API.
Dues to it having a similar feature to another plugin we had I renamed the old plugin to tendysub, as that plugin was only for the chicken tender sub.
Commands that comes with the plugin
[blank] or random: gets a random sub and return the current price the last sale date, and a picture.
[sub name]: gets the current price, the last sale date, and a picture of the sub
list or types: return a list of available subs. (I implemented an auto growing feature here as Abe is working on the API still)
update: This command updates the list of available subs.  (Once again implemented as Abe is still working on the API, and I wanted to keep the idea that the list only needs to be updated from time to time)